+++ src/main/webapp/WEB-INF/data/changelog.md
@@ -1,6 +1,9 @@
 ## 2.1.0
 
 * Upgrade to Grails 3
+* Positive and negative assessments of organizations and persons
+* Boilerplates for inclusion in text controls
+* Lot of bug fixes and code optimizations
 
 ## 2.0.20